attempt to do the heavy duty refactoring
