+++ src/pyved_engine/xecs/_ecs_pattern.py
